pyproject.toml
# This file is used to configure the Python project and its dependencies.

# Build system requirements
[build-system]
requires = ["setuptools>=61"]

# Project metadata
[project]
name = "seamcarver"
version = "0.2.0"
description = "A command-line Python tool for image resizing using seam carving."
authors = [{ name = "Albert Okine", email = "okinealb@grinnell.edu" }]
license = { file = "LICENSE" }
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.8"
dependencies = [
    "numpy",
    "Pillow",
    "scipy",
]
classifiers = [
    # Project classification
    "Development Status :: 3 - Pre-Alpha",
    # Audience
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Developers",
    # Environment
    "Environment :: Console",
    # License
    "License :: OSI Approved :: MIT License",
    # OS Compatibility
    "Operating System :: OS Independent",
    # Programming Language
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    # Topics
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Utilities",
]

# URLs for the main project repository
[project.urls]
Repository = "https://github.com/okinealb/seamcarver.git"

# Optional scripts to run the command-line interface
[project.scripts]
seamcarver = "seamcarver.cli:main"

# Optional dependencies for development and testing
[project.optional-dependencies]
test = [
    "pytest",
    "pytest-benchmark",
]

# Configuration for setuptools to exclude testing packages
[tool.setuptools.packages.find]
exclude = ["tests*", "test*", "build*"]